backend/requirements.txt
flask
flask-cors
transformers
sentence-transformers
torch
scikit-learn
langchain
langchain-community
langchain-huggingface
google-generativeai
python-dotenv
